requirements.txt
django==1.10.7
djangorestframework==3.6.3